Move params out of lib.rs
